PyQt5-Starting-Basic-Project
This code is a Python application that performs interface development using the PyQt5 library. This application contains a window, a text field, a "Clear" button and a "Write" button. Clicking the "Clear" button clears the data in the text area, while clicking the "Write" button prints the data in the text area.

In addition, the application includes features such as background color and adding images.

This code can be used as a basic example of UI development. However, more complex applications can be developed by customizing it according to user needs.